requirements.txt
fastapi
uvicorn
pyyaml
redis
sqlalchemy
psycopg2-binary
pandas
python-dotenv==0.19.0
mlflow==1.20.1
sqlalchemy-views==0.3.0